Move UniqueCardList back to powercard module
