+++ src/main/java/seedu/address/model/powercard/UniqueCardList.java
@@ -1,4 +1,4 @@
-package seedu.address.model.powerdeck;
+package seedu.address.model.powercard;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
@@ -10,7 +10,6 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import seedu.address.model.person.exceptions.DuplicatePersonException;
 import seedu.address.model.person.exceptions.PersonNotFoundException;
-import seedu.address.model.powercard.PowerCard;
 
 /**
  * A list of cards that enforces uniqueness between its elements and does not allow nulls.
